requirements.txt
streamlit==1.43.0
google.generativeai==0.8.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
passlib==1.7.4
pyperclip==1.8.2
matplotlib==3.10.1